+++ library/include/hiprand/hiprand.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2023 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -31,17 +31,17 @@
 
 /// \cond HIPRAND_DOCS_MACRO
 #ifndef HIPRANDAPI
-#ifdef WIN32
-  #ifdef hiprand_EXPORTS
-  /* We are building this library */
-    #define HIPRANDAPI __declspec(dllexport)
-  #else
-    /* We are using this library */
-    #define HIPRANDAPI __declspec(dllimport)
-  #endif
-#else
-  #define HIPRANDAPI
-#endif
+    #ifdef _WIN32
+        #ifdef hiprand_EXPORTS
+            /* We are building this library */
+            #define HIPRANDAPI __declspec(dllexport)
+        #else
+            /* We are using this library */
+            #define HIPRANDAPI __declspec(dllimport)
+        #endif
+    #else
+        #define HIPRANDAPI
+    #endif
 #endif
 /// \endcond
 
@@ -62,10 +62,10 @@
 #define HIPRAND_VERSION
 #endif
 
-#if defined(__HIP_PLATFORM_HCC__) || defined(__HIP_PLATFORM_AMD__)
-#include "hiprand/hiprand_hcc.h"
+#if defined(__HIP_PLATFORM_AMD__)
+    #include "hiprand/hiprand_rocm.h"
 #elif defined(__HIP_PLATFORM_NVCC__)
-#include "hiprand/hiprand_nvcc.h"
+    #include "hiprand/hiprand_nvcc.h"
 #endif
 
 /// \cond HIPRAND_DOCS_TYPEDEFS
@@ -132,6 +132,13 @@ typedef enum hiprandRngType {
     HIPRAND_RNG_QUASI_SCRAMBLED_SOBOL64 = 504  ///< Scrambled Sobol64 quasirandom generator
 } hiprandRngType_t;
 
+typedef enum hiprandDirectionVectorSet
+{
+    HIPRAND_DIRECTION_VECTORS_32_JOEKUO6           = 101,
+    HIPRAND_SCRAMBLED_DIRECTION_VECTORS_32_JOEKUO6 = 102,
+    HIPRAND_DIRECTION_VECTORS_64_JOEKUO6           = 103,
+    HIPRAND_SCRAMBLED_DIRECTION_VECTORS_64_JOEKUO6 = 104,
+} hiprandDirectionVectorSet_t;
 
 // Host API functions
 
@@ -712,10 +719,20 @@ hiprandCreatePoissonDistribution(double lambda, hiprandDiscreteDistribution_t * 
 hiprandStatus_t HIPRANDAPI
 hiprandDestroyDistribution(hiprandDiscreteDistribution_t discrete_distribution);
 
+hiprandStatus_t HIPRANDAPI hiprandGetDirectionVectors32(hiprandDirectionVectors32_t** vectors,
+                                                        hiprandDirectionVectorSet_t   set);
+
+hiprandStatus_t HIPRANDAPI hiprandGetDirectionVectors64(hiprandDirectionVectors64_t** vectors,
+                                                        hiprandDirectionVectorSet_t   set);
+
+hiprandStatus_t HIPRANDAPI hiprandGetScrambleConstants32(const unsigned int** constants);
+
+hiprandStatus_t HIPRANDAPI hiprandGetScrambleConstants64(const unsigned long long** constants);
+
 #if defined(__cplusplus)
 }
 #endif /* __cplusplus */
 
-#endif // HIPRAND_H_
-
 /** @} */ // end of group hiprandhost
+
+#endif // HIPRAND_H_